feat(configs): enhance install script and add zsh aliases

+++ configs/install.sh
@@ -1,19 +1,35 @@
 #!/bin/bash
 
 DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+SSH_DIR="${HOME}/.ssh"
+GH_CONFIG_DIR="${HOME}/.config/gh"
 
 # create symbolic links
 create_symlink() {
-  ln -sf "${DIR}/$1" "${HOME}/$2"
+  if [ -e "${HOME}/$2" ]; then
+    echo "-> Skipping $2, already exists."
+  else
+    ln -sf "${DIR}/$1" "${HOME}/$2"
+    echo "-> Symlink created for $1"
+  fi
 }
 
 # configure ssh
 configure_ssh() {
-  ssh-keygen -t ed25519 -C "phat@onroads.xyz" -f "${HOME}/.ssh/id_ed25519" -P ""
-  ssh-keygen -t ed25519 -C "phatttpham9@gmail.com" -f "${HOME}/.ssh/id_ed25519_secondary" -P ""
+  mkdir -p "${SSH_DIR}"
+  if [ ! -f "${SSH_DIR}/id_ed25519" ]; then
+    ssh-keygen -t ed25519 -C "phat@onroads.xyz" -f "${SSH_DIR}/id_ed25519" -P ""
+  else
+    echo "-> SSH key id_ed25519 already exists."
+  fi
+  if [ ! -f "${SSH_DIR}/id_ed25519_secondary" ]; then
+    ssh-keygen -t ed25519 -C "phatttpham9@gmail.com" -f "${SSH_DIR}/id_ed25519_secondary" -P ""
+  else
+    echo "-> SSH key id_ed25519_secondary already exists."
+  fi
   create_symlink "ssh/config" ".ssh/config"
   create_symlink "ssh/config_secondary" ".ssh/config_secondary"
-  echo "-> ssh key generated & configured!"
+  echo "-> SSH keys generated & configured!"
   echo "-> Attention! Don't forget to copy the public key & add it to remote hosts such as github.com & gitlab.com."
 }
 
@@ -22,30 +38,23 @@ configure_git() {
   create_symlink "git/gitconfig" ".gitconfig"
   create_symlink "git/gitconfig_secondary" ".gitconfig_secondary"
   create_symlink "git/gitignore_global" ".gitignore_global"
-  echo "-> git configured!"
+  echo "-> Git configured!"
 }
 
 # configure gh
 configure_gh() {
-  if [ ! -d "${HOME}/.config/gh" ]; then
-    mkdir -p "${HOME}/.config/gh"
-  fi
+  mkdir -p "${GH_CONFIG_DIR}"
   create_symlink "gh/config.yml" ".config/gh/config.yml"
   create_symlink "gh/hosts.yml" ".config/gh/hosts.yml"
-  echo "-> gh configured!"
+  echo "-> GitHub CLI configured!"
 }
 
 # configure zsh
 configure_zsh() {
   create_symlink "zsh/zshrc" ".zshrc"
-  echo "-> zsh configured!"
+  create_symlink "zsh/zsh_aliases" ".zsh_aliases"
+  echo "-> Zsh configured!"
 }
-
-# call the configuration functions
-configure_ssh
-configure_git
-configure_gh
-configure_zsh
 
 # install gh extensions
 install_gh_extensions() {
@@ -55,5 +64,9 @@ install_gh_extensions() {
   echo "-> gh extensions installed!"
 }
 
-# call the extension installation functions
+# Main script execution
+configure_ssh
+configure_git
+configure_gh
+configure_zsh
 install_gh_extensions
